+++ CHANGELOG.md
@@ -9,11 +9,13 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Add [`backgroundColor`](https://threepipe.org/docs/interfaces/ThreeViewerOptions.html#backgroundColor) to `ThreeViewerOptions` to set initial background color of the scene. Note that background map can be set by `load.background` in the options.
+- Add `uiconfig` in `RootScene` to toggle transparent background.
+- Add `backgroundTonemap` in `RootScene` to toggle tonemapping of background color and map.
 
 ### Changed
 
 - Set `colorSpace` of texture returned by [dataTextureFromVec4](https://threepipe.org/docs/functions/dataTextureFromVec4.html) to `LinearSRGBColorSpace` to be consistent with other texture creation functions.
-
+- Remove `uiconfig` for `tonemapBackground` in `TonemapPlugin` in favor of `backgroundTonemap` UI in `RootScene`.
 
 ## [0.2.0] - 2025-09-03